Further (>10%) speed improvements
- switched to faster amd64 assembly implementation of sha512
- changed the hashset implementation from 8 byte  to 4 byte pointers.
- switch seed generation to 128bit integer intrinsics
-

+++ hashset/hasher.h
@@ -1,7 +1,8 @@
 #ifndef HASHER_H
 #define HASHER_H
 
-#include <stddef.h>
-size_t hasher(const void *element, size_t len);
+#include <stdint.h>
+
+uint32_t hasher(const void *element, uint32_t len);
 
 #endif
